requirements.txt
cairocffi===file-.cairocffi-VERSION
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
idna==2.10
lottie==0.6.7
numpy==1.19.5
opencv-python==4.5.2.54
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
pyTelegramBotAPI==3.7.9
requests==2.25.1
tinycss2==1.1.0
urllib3==1.26.5
webencodings==0.5.1
